Updated links
Not sure if I'm just using it wrong, but from my end, the links return a 404 error. I directly added the full paths to the documentation links. For the Input_File link, I linked to the "Events and Malfunctions" heading, since that would seem to be the most useful section for someone coming from the Event-Manager page.